SciGenML

Documentation for SciGenML.

SciGenML.Config.ConditionalFlowMatchingHyperparameters

SciGenML.Config.DenseNeuralNetworkHyperparameters

SciGenML.Config.FlowMatchingHyperparameters

SciGenML.Config.Hyperparameters

SciGenML.Config.OptimizerHyperparameters

SciGenML.Config.ScoreBasedDiffusionModelHyperparameters

SciGenML.Config.StochasticInterpolantHyperparameters

SciGenML.Config.TrainingHyperparameters

SciGenML.Config.UNetHyperparameters

SciGenML.Models.ConditionalFlowMatching

SciGenML.Models.ConditionalGenerativeModel

SciGenML.Models.DeterministicInterpolantCoefs

SciGenML.Models.FlowMatching

SciGenML.Models.GenerativeModel

SciGenML.Models.ScoreBasedDiffusionModel

SciGenML.Models.StochasticInterpolant

SciGenML.Models.StochasticInterpolantCoefs

SciGenML.Architectures.DenseNeuralNetwork

SciGenML.Architectures.UNet

SciGenML.Architectures.UNet

SciGenML.Architectures.get_model

SciGenML.Layers.conv_next_block

SciGenML.Layers.conv_next_block_no_conditioning

SciGenML.Layers.get_padding

SciGenML.Layers.multiple_conv_next_blocks

SciGenML.Layers.multiple_conv_next_blocks_no_conditioning

SciGenML.Layers.sinusoidal_embedding

SciGenML.Models.compute_interpolant

SciGenML.Models.compute_interpolant

SciGenML.Models.compute_interpolant_diff

SciGenML.Models.diffusion_interpolant_coefs

SciGenML.Models.drift_term

SciGenML.Models.drift_term

SciGenML.Models.drift_term

SciGenML.Models.drift_term

SciGenML.Models.drift_term

SciGenML.Models.get_interpolant_coefs

SciGenML.Models.linear_interpolant_coefs

SciGenML.Models.quadratic_interpolant_coefs

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.TimeIntegrators.RK4_step

SciGenML.TimeIntegrators.RK4_step

SciGenML.TimeIntegrators.euler_maruyama_step

SciGenML.TimeIntegrators.euler_maruyama_step

SciGenML.TimeIntegrators.forward_euler_step

SciGenML.TimeIntegrators.forward_euler_step

SciGenML.TimeIntegrators.heun_step

SciGenML.TimeIntegrators.heun_step

SciGenML.TimeIntegrators.ode_integrator

SciGenML.TimeIntegrators.ode_integrator

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training.compute_score_loss

SciGenML.Training.compute_velocity_loss

SciGenML.Training.get_dataloader

SciGenML.Training.get_dataloader

SciGenML.Training.get_dataloader

SciGenML.Training.get_gradients

SciGenML.Training.get_interpolated_samples

SciGenML.Training.get_interpolated_samples

SciGenML.Training.get_optimizer

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Models.ConditionalFlowMatching — Type

ConditionalFlowMatching

A conditional flow matching generative model.

The conditional flow matching generative model is a generative model that uses a flow matching approach to generate data.

It is a special case of the conditional stochastic interpolant generative model where there is no noise in the interpolant.

SciGenML.Models.ConditionalGenerativeModel — Type

Abstract type for all conditional generative models.

SciGenML.Models.DeterministicInterpolantCoefs — Type

DeterministicInterpolantCoefs

A struct that contains the alpha and beta function coefficients and their derivatives.

The interpolant is defined as:
x(t) = α(t) * x0 + β(t) * x1

The derivatives are:
dx/dt = α'(t) * x0 + β'(t) * x1

SciGenML.Models.FlowMatching — Type

FlowMatching

A flow matching generative model.

The flow matching generative model is a generative model that uses a flow matching approach to generate data.

It is a special case of the stochastic interpolant generative model where there is no noise in the interpolant.

SciGenML.Models.GenerativeModel — Type

Abstract type for all generative models.

SciGenML.Models.ScoreBasedDiffusionModel — Type

ScoreBasedDiffusionModel

A score-based diffusion model generative model.

The score-based diffusion model generative model is a generative model that uses a score-based diffusion model approach to generate data.

It is a special case of the stochastic interpolant generative model where there is no noise in the interpolant.

SciGenML.Models.StochasticInterpolant — Type

StochasticInterpolant

A stochastic interpolant generative model.

SciGenML.Models.StochasticInterpolantCoefs — Type

StochasticInterpolantCoefs

A struct that contains the alpha and beta function coefficients and their derivatives.

The interpolant is defined as:
x(t) = α(t) * x0 + β(t) * x1 + γ(t) * Z

The derivatives are:
dx/dt = α'(t) * x0 + β'(t) * x1 + γ'(t) * Z

SciGenML.Models.compute_interpolant — Method

compute_interpolant(x0, x1, z, t, interpolant_coefs::StochasticInterpolantCoefs)

Computes the stochastic interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    z: The noise.
    t: The time.
    interpolant_coefs: The interpolant coefficients.

SciGenML.Models.compute_interpolant — Method

Computes the deterministic interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    t: The time.
    interpolant_coefs: The interpolant coefficients.

SciGenML.Models.compute_interpolant_diff — Method

compute_interpolant_diff

Computes the derivative of the interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    interpolant_coefs: The interpolant coefficients.
    t: The time.

SciGenML.Models.diffusion_interpolant_coefs — Function

diffusion_interpolant_coefs(trait::Models.Deterministic)

Returns the quadratic interpolant.

The quadratic interpolant is defined as:
α(t) = exp(-multiplier * t)
β(t) = sqrt(1 - exp(-2 * multiplier * t))

The derivatives are:
α'(t) = -multiplier * exp(-multiplier * t)
β'(t) = multiplier * exp(-multiplier * t) / sqrt(1 - exp(-2 * multiplier * t))

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::Models.ConditionalFlowMatching,
)

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Deterministic,
    model::Models.ConditionalFlowMatching,
)

Get the drift term for the conditional flow matching generative model.

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::StochasticInterpolant, 
    diffusion_fn, 
)

Compute the drift term for a stochsatic interpolant.

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::StochasticInterpolant, 
    diffusion_fn, 
)

Compute the drift term for a stochsatic interpolant.

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::StochasticInterpolant, 
    diffusion_fn, 
    x, 
    ps, 
    st
)

Compute the drift term for a stochastic interpolant.

SciGenML.Models.get_interpolant_coefs — Method

get_interpolant_coefs(type::String)

Get the interpolant coefficients for a given type. 
Supported types are:
- "linear"
- "quadratic"

SciGenML.Models.linear_interpolant_coefs — Function

linear_interpolant_coefs(trait::Models.Deterministic)

Returns the linear interpolant.

The linear interpolant is defined as:
α(t) = 1 - t
β(t) = t
γ(t) = 1 - t, if trait == Models.Stochastic

The derivatives are:
α'(t) = -1
β'(t) = 1
γ'(t) = -1, if trait == Models.Stochastic

SciGenML.Models.quadratic_interpolant_coefs — Function

quadratic_interpolant_coefs(trait::Models.Deterministic)

Returns the quadratic interpolant.

The quadratic interpolant is defined as:
α(t) = 1 - t
β(t) = t^2
γ(t) = 1 - t, if trait == Models.Stochastic

The derivatives are:
α'(t) = -1
β'(t) = 2t
γ'(t) = -1, if trait == Models.Stochastic

SciGenML.Architectures.DenseNeuralNetwork — Method

DenseNN(
    in_features::Tuple,
    out_features::Int, 
    hidden_features::Vector{Int};
    activation_function::Function = DEFAULT_ACTIVATION_FUNCTION,
    dropout::Float32 = DEFAULT_DROPOUT
)

in_features: A tuple of the number of input features
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.

A dense neural network with `in_features` input features, 
`out_features` output features, and `hidden_features` hidden features.
The input is a tuple that are concatenated to form a single input.

SciGenML.Architectures.UNet — Method

UNet(
    in_channels::Int,
    out_channels::Int,
    hidden_channels,
    time_embedding_dim::Int
    padding::String,
)

A UNet model.

# Arguments
- `in_channels::Int`: The number of input channels.
- `out_channels::Int`: The number of output channels.
- `hidden_channels::List{Int}`: The number of hidden channels in each layer.
- `time_embedding_dim::Int`: The dimension of the time embedding.
- `padding::String`: The padding type.

SciGenML.Architectures.UNet — Method

UNet(
    in_channels::Int,
    out_channels::Int,
    hidden_channels,
    time_embedding_dim::Int
    padding::String,
)

A UNet model.

# Arguments
- `in_channels::Int`: The number of input channels.
- `out_channels::Int`: The number of output channels.
- `hidden_channels::List{Int}`: The number of hidden channels in each layer.
- `time_embedding_dim::Int`: The dimension of the time embedding.
- `padding::String`: The padding type.

SciGenML.Architectures.get_model — Method

get_model(
    in_features,
    out_features,
    hidden_features,
    activation_function = DEFAULT_ACTIVATION_FUNCTION,
    dropout = DEFAULT_DROPOUT
)

in_features: The number of input features.
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.
dropout: The dropout rate.

Returns a model that can be used to predict the output from the input.

SciGenML.Layers.conv_next_block — Method

conv_next_block(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    pars_embed_dim::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a conv next block with the given number of input and output channels. The block consists of two convolutional layers with kernel size kernel_size. The first layer has the same number of input and output channels, while the second layer has the same number of output channels as the block. The block also includes batch normalization and a skip connection.

https://arxiv.org/abs/2201.03545

Based on https://github.com/tum-pbs/autoreg-pde-diffusion/blob/b9b33913b99ede88d9452c5ab470c5d7f5da5c56/src/turbpred/modeldiffusionblocks.py#L60

SciGenML.Layers.conv_next_block_no_conditioning — Method

conv_next_block_no_conditioning(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a conv next block with the given number of input and output channels. The block consists of two convolutional layers with kernel size kernel_size. The first layer has the same number of input and output channels, while the second layer has the same number of output channels as the block. The block also includes batch normalization and a skip connection.

https://arxiv.org/abs/2201.03545

Based on https://github.com/tum-pbs/autoreg-pde-diffusion/blob/b9b33913b99ede88d9452c5ab470c5d7f5da5c56/src/turbpred/modeldiffusionblocks.py#L60

SciGenML.Layers.get_padding — Method

get_padding(padding::String, padding_size::Int)

Get the padding function for the given padding type and size.

SciGenML.Layers.multiple_conv_next_blocks — Method

multiple_conv_next_blocks(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    conditioning_dim::Int = 1,
    imsize::Tuple{Int, Int} = (64, 128)
)

Create a chain of two conv next blocks with the given number of input and output channels. The first block has the same number of input and output

SciGenML.Layers.multiple_conv_next_blocks_no_conditioning — Method

multiple_conv_next_blocks_no_conditioning(
    in_channels::Int, 
    out_channels::Int,
    multiplier::Int = 1,
    padding="constant"
)

Create a chain of two conv next blocks with the given number of input and output channels. The first block has the same number of input and output

SciGenML.Layers.sinusoidal_embedding — Function

sinusoidal_embedding(
    x::AbstractArray{AbstractFloat, 4},
    min_freq::AbstractFloat,
    max_freq::AbstractFloat,
    embedding_dims::Int
)

Embed the noise variances to a sinusoidal embedding with the given frequency range and embedding dimensions.

Based on https://yng87.page/en/blog/2022/lux-ddim/.

SciGenML.TimeIntegrators.RK4_step — Method

RK4_step(rhs, x, t, dt, ps, st)

Perform a 4th order Runge-Kutta step.

SciGenML.TimeIntegrators.RK4_step — Method

RK4_step(rhs, x, scalar_conditioning, t, dt, ps, st)

Perform a 4th order Runge-Kutta step.

SciGenML.TimeIntegrators.euler_maruyama_step — Method

Euler Maruyama step for SDEs.

SciGenML.TimeIntegrators.euler_maruyama_step — Method

Euler Maruyama step for SDEs.

SciGenML.TimeIntegrators.forward_euler_step — Method

forward_euler_step(model, x, t, dt)

Perform a forward Euler step.

SciGenML.TimeIntegrators.forward_euler_step — Method

forward_euler_step(model, x, scalar_conditioning, t, dt)

Perform a forward Euler step.

SciGenML.TimeIntegrators.heun_step — Method

Heun step for SDEs.

SciGenML.TimeIntegrators.heun_step — Method

Heun step for SDEs.

SciGenML.TimeIntegrators.ode_integrator — Method

ode_integrator(
    stepper, 
    rhs, 
    x, 
    num_steps, 
    ps, 
    st;
    t_interval=[0.0, 1.0], 
    verbose::Bool = true
)

Generic ODE integrator.

SciGenML.TimeIntegrators.ode_integrator — Method

ode_integrator(
    stepper, 
    rhs, 
    x, 
    scalar_conditioning,
    num_steps, 
    ps, 
    st;
    t_interval=[0.0, 1.0], 
    verbose::Bool = true
)

Generic ODE integrator.

SciGenML.Sampling.sample — Method

sample(
    model,
    args...; 
    kwargs...
)

Generic training function for all models.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.ScoreBasedDiffusionModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a denoising diffusion model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.StochasticInterpolantGenerativeModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Stochastic,
    model::Models.ScoreBasedDiffusionModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a denoising diffusion model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Stochastic,
    model::Models.ScoreBasedDiffusionModel,
    num_samples::Int,
    num_steps::Int,
    num_samples::Int,
    prior_samples,
    diffusion_fn,
    rng::Random.AbstractRNG = Random.default_rng(),
    verbose::Bool = true,
    stepper = TimeIntegrators.heun_step
)

Sample from a denoising diffusion model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Stochastic,
    model::Models.StochasticInterpolant,
    scalar_conditioning,
    num_steps::Int,
    num_samples::Int,
    prior_samples = nothing,
    diffusion_fn = nothing,
    rng::Random.AbstractRNG = Random.default_rng(),
    verbose::Bool = true,
    stepper = TimeIntegrators.heun_step
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Stochastic,
    model::Models.StochasticInterpolantGenerativeModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.ConditionalFlowMatching,
    base_samples,
    target_samples,
    scalar_conditioning_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.StochasticInterpolant,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.ScoreBasedDiffusionModel,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a score-based diffusion generative model for one step via flow matching.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.ScoreBasedDiffusionModel,
    base_samples,
    target_samples,
    scalar_conditioning,
    train_state,
    rng
)

Train a score-based diffusion generative model for one step via flow matching.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Stochastic,
    model::Models.StochasticInterpolant,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Stochastic,
    model::Models.StochasticInterpolant,
    base_samples,
    target_samples,
    scalar_conditioning,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training.compute_score_loss — Method

compute_score_loss(model, ps, st, (x, z, gamma))

Compute the loss for a stochastic interpolant generative model.

SciGenML.Training.compute_velocity_loss — Method

compute_velocity_loss(model, ps, st, (x, y))

Compute the loss for a stochastic interpolant generative model.

SciGenML.Training.get_dataloader — Function

get_dataloader(
    base, 
    target, 
    batch_size::Int, 
    match_base_and_target::Bool=false
)

Get a dataloader from a base and target.

SciGenML.Training.get_dataloader — Function

get_dataloader(
    data::NamedTuple,
    batch_size::Int,
    match_base_and_target::Bool=false
)

Get a dataloader from a named tuple.

SciGenML.Training.get_dataloader — Function

get_dataloader(
    base,
    target,
    scalar_conditioning,
    batch_size::Int,
    match_base_and_target::Bool=false
)

Get a dataloader from a base, target, and conditioning.

SciGenML.Training.get_gradients — Method

get_gradients(
    model,
    batch,
    train_state,
    loss_fn,
)

Get gradients from a model, batch, train state, loss function, and random number generator.

SciGenML.Training.get_interpolated_samples — Method

get_interpolated_samples(
    base_samples,
    target_samples,
    t_samples,
    interpolant_coefs
)

Get interpolated samples from a base, target, and time.

SciGenML.Training.get_interpolated_samples — Method

get_interpolated_samples(
    base_samples,
    target_samples,
    z_samples,
    t_samples,
    interpolant_coefs

)

Get interpolated samples from a base, target, and time.

SciGenML.Training.get_optimizer — Method

get_optimizer(type::String, lr::Float32, lambda::Float32)

Get an optimizer from a string.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.ConditionalFlowMatching,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Stochastic,
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.ScoreBasedDiffusionModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a score-based diffusion generative model vis flow matching.

SciGenML.Training.train — Method

train(
    model,
    args...; 
    kwargs...
)

Generic training function for all models.

SciGenML.Config.ConditionalFlowMatchingHyperparameters — Type

ConditionalFlowMatchingHyperparameters

Hyperparameters for the conditional flow matching generative model.

SciGenML.Config.DenseNeuralNetworkHyperparameters — Type

DenseNeuralNetworkHyperparameters

Hyperparameters for the dense neural network.

SciGenML.Config.FlowMatchingHyperparameters — Type

FlowMatchingHyperparameters

Hyperparameters for the flow matching generative model.

SciGenML.Config.Hyperparameters — Type

Hyperparameters

Hyperparameters for the architecture, training, and optimizer.

SciGenML.Config.OptimizerHyperparameters — Type

OptimizerHyperparameters

Hyperparameters for the optimizer.

SciGenML.Config.ScoreBasedDiffusionModelHyperparameters — Type

ScoreBasedDiffusionModelHyperparameters

Hyperparameters for the score-based diffusion model generative model.

SciGenML.Config.StochasticInterpolantHyperparameters — Type

StochasticInterpolantHyperparameters

Hyperparameters for the stochastic interpolant generative model.

SciGenML.Config.TrainingHyperparameters — Type

TrainingHyperparameters

Hyperparameters for the training.

SciGenML.Config.UNetHyperparameters — Type

UNetHyperparameters

Hyperparameters for the UNet.