SciGenML

Documentation for SciGenML.

SciGenML.Config.DenseNeuralNetworkHyperparameters

SciGenML.Config.Hyperparameters

SciGenML.Config.OptimizerHyperparameters

SciGenML.Config.PlaceholderGenerativeModel

SciGenML.Config.PlaceholderHyperparameters

SciGenML.Config.StochasticInterpolantGenerativeModelHyperparameters

SciGenML.Config.TrainingHyperparameters

SciGenML.Models.ConditionalGenerativeModel

SciGenML.Models.GenerativeModel

SciGenML.Models.InterpolantCoefs

SciGenML.Models.StochasticInterpolantGenerativeModel

SciGenML.Architectures.DenseNeuralNetwork

SciGenML.Architectures.get_model

SciGenML.Models.compute_interpolant

SciGenML.Models.compute_interpolant_diff

SciGenML.Models.get_interpolant_coefs

SciGenML.Models.linear_interpolant_coefs

SciGenML.Models.quadratic_interpolant_coefs

SciGenML.Sampling.sample

SciGenML.Training.compute_loss

SciGenML.Training.get_optimizer

SciGenML.Training.prepare_batches

SciGenML.Training.stochastic_interpolant_train_step

SciGenML.Training.train

SciGenML.Models.ConditionalGenerativeModel — Type

Abstract type for all conditional generative models.

SciGenML.Models.GenerativeModel — Type

Abstract type for all generative models.

SciGenML.Models.InterpolantCoefs — Type

InterpolantCoefs

A struct that contains the alpha and beta function coefficients and their derivatives.

The interpolant is defined as:
x(t) = α(t) * x0 + β(t) * x1

The derivatives are:
dx/dt = α'(t) * x0 + β'(t) * x1

SciGenML.Models.StochasticInterpolantGenerativeModel — Type

StochasticInterpolantGenerativeModel

A generative model that uses a stochastic interpolant to generate data.

SciGenML.Models.compute_interpolant — Method

compute_interpolant

Computes the interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    interpolant: The interpolant.
    t: The time.

SciGenML.Models.compute_interpolant_diff — Method

compute_interpolant_diff

Computes the derivative of the interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    interpolant: The interpolant.
    t: The time.

SciGenML.Models.get_interpolant_coefs — Method

get_interpolant_coefs(type::String)

Get the interpolant coefficients for a given type. 
Supported types are:
- "linear"
- "quadratic"

SciGenML.Models.linear_interpolant_coefs — Method

linear_interpolant_coefs

Returns the linear interpolant.

The linear interpolant is defined as:
α(t) = 1 - t
β(t) = t

The derivatives are:
α'(t) = -1
β'(t) = 1

SciGenML.Models.quadratic_interpolant_coefs — Method

quadratic_interpolant_coefs

Returns the quadratic interpolant.

The quadratic interpolant is defined as:
α(t) = 1 - t
β(t) = t^2

The derivatives are:
α'(t) = -1
β'(t) = 2t

SciGenML.Architectures.DenseNeuralNetwork — Method

DenseNN(
    in_features::Tuple,
    out_features::Int, 
    hidden_features::Vector{Int};
    activation_function::Function = DEFAULT_ACTIVATION_FUNCTION,
    dropout::Float32 = DEFAULT_DROPOUT
)

in_features: A tuple of the number of input features
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.

A dense neural network with `in_features` input features, 
`out_features` output features, and `hidden_features` hidden features.
The input is a tuple that are concatenated to form a single input.

SciGenML.Architectures.get_model — Method

get_model(
    in_features,
    out_features,
    hidden_features,
    activation_function = DEFAULT_ACTIVATION_FUNCTION,
    dropout = DEFAULT_DROPOUT
)

in_features: The number of input features.
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.
dropout: The dropout rate.

Returns a model that can be used to predict the output from the input.

SciGenML.Sampling.sample — Method

sample(
    model::Models.StochasticInterpolantGenerativeModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Training.compute_loss — Method

compute_loss(model, ps, st, (x, y))

Compute the loss for a stochastic interpolant generative model.

SciGenML.Training.get_optimizer — Method

get_optimizer(type::String, lr::Float32, lambda::Float32)

Get an optimizer from a string.

SciGenML.Training.prepare_batches — Method

prepare_batches(data::Dict{String, AbstractArray}, batch_size::Int)

Prepare batches for training.

SciGenML.Training.stochastic_interpolant_train_step — Method

stochastic_interpolant_train_step(
    model::Models.StochasticInterpolantGenerativeModel,
    base_samples::AbstractArray,
    target_samples::AbstractArray,
    train_state::Lux.Training.TrainState,
    rng::Random.AbstractRNG
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training.train — Function

train(
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Config.DenseNeuralNetworkHyperparameters — Type

DenseNeuralNetworkHyperparameters

Hyperparameters for the dense neural network.

SciGenML.Config.Hyperparameters — Type

Hyperparameters

Hyperparameters for the architecture, training, and optimizer.

SciGenML.Config.OptimizerHyperparameters — Type

OptimizerHyperparameters

Hyperparameters for the optimizer.

SciGenML.Config.PlaceholderGenerativeModel — Type

PlaceholderGenerativeModel

A generative model that uses a placeholder.

SciGenML.Config.PlaceholderHyperparameters — Type

PlaceholderHyperparameters

Hyperparameters for the placeholder.

SciGenML.Config.StochasticInterpolantGenerativeModelHyperparameters — Type

StochasticInterpolantGenerativeModelHyperparameters

Hyperparameters for the stochastic interpolant generative model.

SciGenML.Config.TrainingHyperparameters — Type

TrainingHyperparameters

Hyperparameters for the training.