SciGenML

Documentation for SciGenML.

SciGenML.Config.DenseNeuralNetworkHyperparameters

SciGenML.Config.FlowMatchingHyperparameters

SciGenML.Config.Hyperparameters

SciGenML.Config.OptimizerHyperparameters

SciGenML.Config.PlaceholderHyperparameters

SciGenML.Config.StochasticInterpolantHyperparameters

SciGenML.Config.TrainingHyperparameters

SciGenML.Models.ConditionalGenerativeModel

SciGenML.Models.DeterministicInterpolantCoefs

SciGenML.Models.FlowMatching

SciGenML.Models.GenerativeModel

SciGenML.Models.StochasticInterpolant

SciGenML.Models.StochasticInterpolantCoefs

SciGenML.Architectures.DenseNeuralNetwork

SciGenML.Architectures.get_model

SciGenML.Models.compute_interpolant

SciGenML.Models.compute_interpolant

SciGenML.Models.compute_interpolant_diff

SciGenML.Models.drift_term

SciGenML.Models.drift_term

SciGenML.Models.get_interpolant_coefs

SciGenML.Models.linear_interpolant_coefs

SciGenML.Models.quadratic_interpolant_coefs

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.Sampling.sample

SciGenML.TimeIntegrators.RK4_step

SciGenML.TimeIntegrators.euler_maruyama_step

SciGenML.TimeIntegrators.forward_euler_step

SciGenML.TimeIntegrators.heun_step

SciGenML.TimeIntegrators.ode_integrator

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training._train_step

SciGenML.Training.compute_score_loss

SciGenML.Training.compute_velocity_loss

SciGenML.Training.get_optimizer

SciGenML.Training.prepare_batches

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Training.train

SciGenML.Models.ConditionalGenerativeModel — Type

Abstract type for all conditional generative models.

SciGenML.Models.DeterministicInterpolantCoefs — Type

DeterministicInterpolantCoefs

A struct that contains the alpha and beta function coefficients and their derivatives.

The interpolant is defined as:
x(t) = α(t) * x0 + β(t) * x1

The derivatives are:
dx/dt = α'(t) * x0 + β'(t) * x1

SciGenML.Models.FlowMatching — Type

FlowMatching

A flow matching generative model.

The flow matching generative model is a generative model that uses a flow matching approach to generate data.

It is a special case of the stochastic interpolant generative model where there is no noise in the interpolant.

SciGenML.Models.GenerativeModel — Type

Abstract type for all generative models.

SciGenML.Models.StochasticInterpolant — Type

StochasticInterpolant

A stochastic interpolant generative model.

SciGenML.Models.StochasticInterpolantCoefs — Type

StochasticInterpolantCoefs

A struct that contains the alpha and beta function coefficients and their derivatives.

The interpolant is defined as:
x(t) = α(t) * x0 + β(t) * x1 + γ(t) * Z

The derivatives are:
dx/dt = α'(t) * x0 + β'(t) * x1 + γ'(t) * Z

SciGenML.Models.compute_interpolant — Method

compute_interpolant(x0, x1, z, t, interpolant_coefs::StochasticInterpolantCoefs)

Computes the stochastic interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    z: The noise.
    t: The time.
    interpolant_coefs: The interpolant coefficients.

SciGenML.Models.compute_interpolant — Method

Computes the deterministic interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    t: The time.
    interpolant_coefs: The interpolant coefficients.

SciGenML.Models.compute_interpolant_diff — Method

compute_interpolant_diff

Computes the derivative of the interpolant at a given time.

Args:
    x0: The starting point.
    x1: The ending point.
    interpolant_coefs: The interpolant coefficients.
    t: The time.

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::StochasticInterpolant, 
    diffusion_fn, 
)

Compute the drift term for a stochsatic interpolant.

SciGenML.Models.drift_term — Method

drift_term(
    ::Models.Stochastic,
    model::StochasticInterpolant, 
    diffusion_fn, 
    x, 
    ps, 
    st
)

Compute the drift term for a stochastic interpolant.

SciGenML.Models.get_interpolant_coefs — Method

get_interpolant_coefs(type::String)

Get the interpolant coefficients for a given type. 
Supported types are:
- "linear"
- "quadratic"

SciGenML.Models.linear_interpolant_coefs — Function

linear_interpolant_coefs(trait::Models.Deterministic)

Returns the linear interpolant.

The linear interpolant is defined as:
α(t) = 1 - t
β(t) = t
γ(t) = 1 - t, if trait == Models.Stochastic

The derivatives are:
α'(t) = -1
β'(t) = 1
γ'(t) = -1, if trait == Models.Stochastic

SciGenML.Models.quadratic_interpolant_coefs — Function

quadratic_interpolant_coefs(trait::Models.Deterministic)

Returns the quadratic interpolant.

The quadratic interpolant is defined as:
α(t) = 1 - t
β(t) = t^2
γ(t) = 1 - t, if trait == Models.Stochastic

The derivatives are:
α'(t) = -1
β'(t) = 2t
γ'(t) = -1, if trait == Models.Stochastic

SciGenML.Architectures.DenseNeuralNetwork — Method

DenseNN(
    in_features::Tuple,
    out_features::Int, 
    hidden_features::Vector{Int};
    activation_function::Function = DEFAULT_ACTIVATION_FUNCTION,
    dropout::Float32 = DEFAULT_DROPOUT
)

in_features: A tuple of the number of input features
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.

A dense neural network with `in_features` input features, 
`out_features` output features, and `hidden_features` hidden features.
The input is a tuple that are concatenated to form a single input.

SciGenML.Architectures.get_model — Method

get_model(
    in_features,
    out_features,
    hidden_features,
    activation_function = DEFAULT_ACTIVATION_FUNCTION,
    dropout = DEFAULT_DROPOUT
)

in_features: The number of input features.
out_features: The number of output features.
hidden_features: The number of hidden features.
activation_function: The activation function.
dropout: The dropout rate.

Returns a model that can be used to predict the output from the input.

SciGenML.TimeIntegrators.RK4_step — Method

RK4_step(rhs, x, t, dt, ps, st)

Perform a 4th order Runge-Kutta step.

SciGenML.TimeIntegrators.euler_maruyama_step — Method

Euler Maruyama step for SDEs.

SciGenML.TimeIntegrators.forward_euler_step — Method

forward_euler_step(model, x, t, dt)

Perform a forward Euler step.

SciGenML.TimeIntegrators.heun_step — Method

Heun step for SDEs.

SciGenML.TimeIntegrators.ode_integrator — Method

ode_integrator(
    stepper, 
    rhs, 
    x, 
    num_steps, 
    ps, 
    st;
    t_interval=[0.0, 1.0], 
    verbose::Bool = true
)

Generic ODE integrator.

SciGenML.Sampling.sample — Method

sample(
    model,
    args...; 
    kwargs...
)

Generic training function for all models.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Deterministic,
    model::Models.StochasticInterpolantGenerativeModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Sampling.sample — Method

sample(
    ::Models.Stochastic,
    model::Models.StochasticInterpolantGenerativeModel,
    num_samples::Int,
    num_steps::Int,
    rng::Random.AbstractRNG = Random.default_rng()
)

Sample from a stochastic interpolant generative model using the forward Euler method.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.FlowMatching,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Deterministic,
    model::Models.StochasticInterpolant,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training._train_step — Method

_train_step(
    ::Models.Stochastic,
    model::Models.StochasticInterpolant,
    base_samples,
    target_samples,
    train_state,
    rng
)

Train a stochastic interpolant generative model for one step.

SciGenML.Training.compute_score_loss — Method

compute_score_loss(model, ps, st, (x, z, gamma))

Compute the loss for a stochastic interpolant generative model.

SciGenML.Training.compute_velocity_loss — Method

compute_velocity_loss(model, ps, st, (x, y))

Compute the loss for a stochastic interpolant generative model.

SciGenML.Training.get_optimizer — Method

get_optimizer(type::String, lr::Float32, lambda::Float32)

Get an optimizer from a string.

SciGenML.Training.prepare_batches — Method

prepare_batches(data::Dict{String, AbstractArray}, batch_size::Int)

Prepare batches for training.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Deterministic,
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Function

train(
    ::Models.Stochastic,
    model::Models.StochasticInterpolantGenerativeModel,
    data,
    config,
    rng = Random.default_rng();
    verbose = true
)

Train a stochastic interpolant generative model.

SciGenML.Training.train — Method

train(
    model,
    args...; 
    kwargs...
)

Generic training function for all models.

SciGenML.Config.DenseNeuralNetworkHyperparameters — Type

DenseNeuralNetworkHyperparameters

Hyperparameters for the dense neural network.

SciGenML.Config.FlowMatchingHyperparameters — Type

FlowMatchingHyperparameters

Hyperparameters for the flow matching generative model.

SciGenML.Config.Hyperparameters — Type

Hyperparameters

Hyperparameters for the architecture, training, and optimizer.

SciGenML.Config.OptimizerHyperparameters — Type

OptimizerHyperparameters

Hyperparameters for the optimizer.

SciGenML.Config.PlaceholderHyperparameters — Type

PlaceholderHyperparameters

Hyperparameters for the placeholder.

SciGenML.Config.StochasticInterpolantHyperparameters — Type

StochasticInterpolantHyperparameters

Hyperparameters for the stochastic interpolant generative model.

SciGenML.Config.TrainingHyperparameters — Type

TrainingHyperparameters

Hyperparameters for the training.